# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactSubmissionInputReason = typeof ContactSubmissionInputReason[keyof typeof ContactSubmissionInputReason];


export const ContactSubmissionInputReason = {
  Comment: 'Comment',
  Question: 'Question',
  Partnership: 'Partnership',
  Opportunity: 'Opportunity',
  Other: 'Other',
} as const;

export interface ContactSubmissionInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email: string;
  reason: ContactSubmissionInputReason;
  /** @minLength 1 */
  message: string;
}

export type ContactSubmissionReason = typeof ContactSubmissionReason[keyof typeof ContactSubmissionReason];


export const ContactSubmissionReason = {
  Comment: 'Comment',
  Question: 'Question',
  Partnership: 'Partnership',
  Opportunity: 'Opportunity',
  Other: 'Other',
} as const;

export interface ContactSubmission {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  reason: ContactSubmissionReason;
  message: string;
  submittedAt: string;
  replied: boolean;
  /** @nullable */
  repliedAt: string | null;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminLoginResult {
  authenticated: boolean;
}

export type ReasonCountReason = typeof ReasonCountReason[keyof typeof ReasonCountReason];


export const ReasonCountReason = {
  Comment: 'Comment',
  Question: 'Question',
  Partnership: 'Partnership',
  Opportunity: 'Opportunity',
  Other: 'Other',
} as const;

export interface ReasonCount {
  reason: ReasonCountReason;
  count: number;
}

export interface AdminSummary {
  totalMessages: number;
  newMessages: number;
  repliedMessages: number;
  replyRate: number;
  byReason: ReasonCount[];
}

export interface ErrorResponse {
  error: string;
}