# lib/api-zod/src/generated/types/adminSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReasonCount } from './reasonCount';

export interface AdminSummary {
  totalMessages: number;
  newMessages: number;
  repliedMessages: number;
  replyRate: number;
  byReason: ReasonCount[];
}