# lib/api-zod/src/generated/types/contactSubmission.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactSubmissionReason } from './contactSubmissionReason';

export interface ContactSubmission {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  reason: ContactSubmissionReason;
  message: string;
  submittedAt: Date;
  replied: boolean;
  /** @nullable */
  repliedAt: Date | null;
}