# lib/api-zod/src/generated/types/contactSubmissionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactSubmissionInputReason } from './contactSubmissionInputReason';

export interface ContactSubmissionInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email: string;
  reason: ContactSubmissionInputReason;
  /** @minLength 1 */
  message: string;
}