# lib/api-zod/src/generated/types/reasonCount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReasonCountReason } from './reasonCountReason';

export interface ReasonCount {
  reason: ReasonCountReason;
  count: number;
}